requirements.txt
# Bazaar Tracker runtime + test dependencies.
# Versions pinned to what was verified available locally; the three
# unpinned packages are environment-sensitive (Windows-only venv or
# game-build-dependent) and cannot be pinned from this checkout.

# Log watching (Player.log tail)
watchdog  # unpinned — Windows venv, not importable from this checkout

# HTTP client (card cache, static API)
requests==2.33.1

# PyWebView overlay window
pywebview  # unpinned — Windows venv, not importable from this checkout

# WSGI server for the Flask dashboard/overlay
waitress==3.0.2

# Dashboard + overlay HTTP app
flask==3.1.3

# Mono hook for managed-memory snapshots
frida  # unpinned — must match game build

# Test framework
pytest==9.0.3
